+++ app/src/test/java/io/github/thachillera/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.github.cardsscorekeeper;
+package io.github.thachillera;
 
 import org.junit.Test;
